Create reusable Arrow component
